% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox_layer.R
\name{add_layer}
\alias{add_layer}
\title{Add layer}
\usage{
add_layer(map, js)
}
\arguments{
\item{map}{A map widget object}

\item{js}{The JSON used to create the layer}
}
\description{
Add layer
}
\examples{
\donttest{
token <- "MAPBOX_TOKEN"

## Terrain in San Francisco, USA
js <- mapbox_terrain

mapbox(
  token = token
  , location = c(-122.44, 37.753)
  , zoom = 10
) \%>\%
  add_layer( js )

## 3-D buildings
js <- mapbox_3d_buildings

mapbox(
  token = token
  , location = c(144.96, -37.82)
  , zoom = 15
  , pitch = 45
) \%>\%
  add_layer( js )

## Terain in Cusco, Peru
source <- '{
	"type": "vector",
	"url": "mapbox://mapbox.mapbox-terrain-v2"
}'

id <- 'contours'

js <- mapbox_contours

mapbox(
  token = token
  , location = c(-71.9675, -13.5320)
  , zoom = 10
) \%>\%
  add_source(
    id = id
    , js = source
  ) \%>\%
  add_layer(
    js = contours
  )

}

}
