% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_style.R
\name{add_style}
\alias{add_style}
\title{Add Style}
\usage{
add_style(map, style)
}
\arguments{
\item{map}{A map widget object}

\item{style}{The style to add}
}
\description{
Add Style
}
\examples{
\donttest{
token <- "MAPBOX_TOKEN"

mapbox(token = token) \%>\%
  add_style(style = mapbox_style(5))
}
}
\seealso{
\code{\link{mapbox_style}}
}
