% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox_map.R
\name{mapbox-shiny}
\alias{mapbox-shiny}
\alias{mapboxOutput}
\alias{renderMapbox}
\title{Shiny bindings for mapbox}
\usage{
mapboxOutput(outputId, width = "100\%", height = "400px")

renderMapbox(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a mapbox}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using mapbox within Shiny
applications and interactive Rmd documents.
}
\examples{
\donttest{
library(mapbox)
library(shiny)

token <- "MAPBOX_TOKEN"

ui <- fluidPage(
  mapbox::mapboxOutput("map", height = "800px")
)

server <- function(input, output, session) {
  output$map <- mapbox::renderMapbox({
    mapbox(token = token, zoom = 10,
           location = c(-74.5447, 40.6892),
           style = mapbox_style(2),
           pitch = 20)
  })
}
shinyApp(ui, server)
}

}
