% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox_layer.R
\docType{data}
\name{mapbox_3d_buildings}
\alias{mapbox_3d_buildings}
\title{3D-Building JSON}
\format{An object of class \code{character} of length 1.}
\usage{
mapbox_3d_buildings
}
\description{
3D-Building JSON
}
\keyword{datasets}
