% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox_layer.R
\docType{data}
\name{mapbox_contours}
\alias{mapbox_contours}
\title{Contours JSON}
\format{An object of class \code{character} of length 1.}
\usage{
mapbox_contours
}
\description{
Contours JSON
}
\keyword{datasets}
