% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox_styles.R
\name{mapbox_style}
\alias{mapbox_style}
\title{mapbox_style}
\usage{
mapbox_style(x = 1, owner = NULL, styleid = NULL, print = FALSE,
  optimized = FALSE)
}
\arguments{
\item{x}{Select a predefined style}

\item{owner}{For custom-made styles, include your Mapbox account name}

\item{styleid}{For custom-made styles, include your style ID}

\item{print}{Prints the predefined styles if TRUE. Default is FALSE}

\item{optimized}{If TRUE will append `?optimize=true` at the end of the
style URL. Default is FALSE}
}
\description{
Get a predefined or custom-made Mapbox style
}
\examples{
\donttest{
mapbox_style(print = TRUE)
mapbox_style(4)
mapbox_style(owner = "user", styleid = "style_id")
}
}
