% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox_layer.R
\docType{data}
\name{mapbox_terrain}
\alias{mapbox_terrain}
\title{Terrain JSON}
\format{An object of class \code{character} of length 1.}
\usage{
mapbox_terrain
}
\description{
Terrain JSON
}
\keyword{datasets}
