% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox_package.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{A mapdeck map and a layer to add to it}
}
\description{
Uses the pipe operator (\code{\%>\%}) to chain statements. Useful for adding
layers to a \code{mapdeck} map
}
